Refactor decodeUtf8 (WIP)
